Add a ProcessingRequestProvider for single files
ProcessingScannable is available as a ProcessingRequestProvider
to add processing to concurrent scans but for the majority of use cases it
is too complex. SingleFileProcessingRequest takes a single file path and
provides a ProcessingRequest that maps a single key to a collection containing
the file.

The path to the file can be relative to a base template so can respond to
the visit being changed etc if processing configurations are stored in a
standard location.

Change-Id: Ic8c81997aafa928432bc7b6b6ed8f434a1f9bdfb