[DAQ-3212] call reset on malcolm device when scan finished
This causes malcolm to close the file. Not doing this causes issues
with archiving the files that malcolm is writing to.

This behaviour can be turned off by calling the static method
MalcolmDevice.setResetAfterScan() with the argument 'false'. This is
required to help with debugging scans as calling reset also clears
Malcolm's error state.

Change-Id: I58c5763673943a567b1096fd3c609b1243764fd8