DAQ-3237 Improve reliability of WatchdogShutterTest
Set longer exposure time for specific watchdog tests.

Test failures were common due to concurrent thread timing issues when
running on a busy machine. This overrides the exposure time set in
AbstractAcquisitionTest (0.08s).

Also an NPE is often thrown from the before method for an unknown
reason so ensure that the stack trace is logged to standard out.

Change-Id: Ib31df43d1b8cf99b4dee55b1f220cc79dc8664a1