B18-472 : Add get, set scannable to Spreadsheet view context menu
New context menu items have been added to SpreadSheet view to allow
the value in a cell to be updated using the current position of a
scannable, and to set the position of a scannable using value in
the selected cell. This will work for SampleParameterMotorPosition
parameters, as typically used in SampleParameter settings.

Change-Id: Ic3b5f525b1c4e16f09a6b028f57e03f41c788874