I07-372 Add Malcolm device for live processing of scans in GDA
A task is started in a new thread which reads and processes frames as
they are available on disk.

The completion of the scan is blocked until all frames have been
processed.

The processed data is contributed to the NXdetector created by the
MaloclmDevice where the datasets are added as additional primary fields.

Change-Id: I1b55cdccf77b23a5cff32d4fd07ec1cfeee6d7e1