K11-969 Moved ScanningAcquisitionService on the server as REST service
The ScanningAcquisitionService requires access to the mapping library
and this was a blocking reason to have as REST service in the core
repository as for the ExperimentRestService. A better solution has been
found in moving all the REST infrastructure in the experiment
repository.

- Relocated SpringWebInitializer, ExperimentRestService and
DocumentMapper in the experiment repository
- Implemented a ScanningAcquisitionService and the associated
ScanningAcquisitionRestService
- Implementation of acquisition REST client (see uk.ac.gda.ui.tool.rest
package)
- Improves the ClientRestService#createHttpEntity with support for json
- Refactored ScanningAcquisitionController to use the REST acquisition
service

Change-Id: I168a592f1d8bacbf9723a361f5a7188e7c641223