[DAQ-3249] add metadata entries for NeXusUtils add check groups in test
NexusDataWriter uses static method in the class NeXusUtils to write
some of the same metadata groups that we want to add in the
Diamond Default Nexus Structure. These look up metadata entries from
the Metadata object and only writes the properties it finds.
NexusDataWriterScanTest now sets up the required properties and
checks that are written to the Nexus file.

Change-Id: If9eab3c9ba95f882a64bd34c3ac1bc3db66cdf82