DAQ-3237 Use a unique dir name per instance of MockFilePathService
The returned "unique" File instances are provided based on the condition
that the File did not exist at the time the check was made however
this is not robust when multiple instances of the test suite are
running on the same machine as there will be a delay between obtaining
a File and creating it i.e. the same File may be returned by two
different processes.

NexusTest must also use the filePathService to be consistent.

Change-Id: I3a949812efdc7912225a54b75a90a0ac6c561d01