DAQ-3276 Show Jython commands from other clients
With multiple clients open, commands from other users are not shown
even though output is. This can make it hard to follow what's going on.

This shows commands from other clients on a client's JythonTerminalView
and in the SSH JythonShell. Commands are prefixed by the username and
index of the client they originated from, eg

    (qan22331@2) >>> scan x 0 10 1 detector_y 3

to make it clear to the user which commands are being run from
elsewhere. The client # is the index used in the BatonManager view if
details are required.

This is optional and can be enabled or disabled via
    Preferences > GDA Preferences > "Show input from all clients..."

Change-Id: I87ef247200af88f8c9e04a917f24c614c7fbd955