DAQ-3114 Move ResponseConfiguration functionality into RequesterImpl
and an enum & constants into the corresponding IRequester interface

The only place (apart from RequesterImpl) that used
ResponseConfiguration was a test in DeviceRequestTest, but this was
deactivated (the @Test was commented out).
I have commented all the code of this test out pending a review of it.

Also updated & tidied the Javadoc

Change-Id: I3431a54a413b0997467162a512877a4e737d59dd