K11-969 Add a new constructor and remove @Value annotation
The only class with a Spring annotation to extend this class is the
ExperimentStructureJobResponder. During the bean creation the
constructor failed because the "@Value broker" was not properly set as
during the @PostConstruct the @Value still has to be processed.
Consequently this change removes the @Value and allows the
ExperimentStructureJobResponder to inject the broker using a dedicated
constructor.

Change-Id: I7e057ddecefdd60d0a4bd85f4b49d2be04a9b426