DAQ-3199 Add beamline health condition to check for ISPyB
This currently uses the SpyCat interface to ISPyB, which is used by
all beamlines to get visit information. The newer gda-ispyb-api library
is used by a few beamlines, but not all have the credentials to access
it.

To support this:
- modify the base class for health conditions extend
FindableConfigurableBase, so we have access to a bean name for
messaging purposes
- factor the rate-control code out of the ActiveMQ condition into
a base class that the ISPyB condition can also inherit from

Change-Id: Ifd9087041b86c13e2614e6d574dd1eb6c599e18f