Don't try and run incomplete Jython commands
When writing interactive commands, all intermediate states are run
to check if they are complete. This moves the check into the JythonServer
so partial commands are not recorded as having been run.

This is particularly useful when moving towards forwarding input
to other clients. Previously, each line would cause the whole command
so far to be echoed to every other client.

Change-Id: Ib0e3d74ec58678933a6fb86be5fd3d9868602768