K11-1069 Fix client crash after second automated experiment
I have not been able to identify the reason for the client crashes, but
confirmed that they were somehow associated with the plotting of the
"plan end" annotation. Therefore this annotation is now plotted in the
same way as the others which are not problematic.

In this change we also lose the 'status' text box at the bottom of the
plot which is now made redundant by the Plan Overview view.

Change-Id: I658a561798ac53b04949cc2c83c395091d004e42