Add utility to listen to json activeMQ messages
ActiveMQ is becoming more widely used for broadcasting messages but
there is no central class for listening for incoming messages and
dispatching them to where they can be processed.

This provides that listener utility. Messages are expected to be JSON
and deserializable to an instance of certain class.  A handler can be
set to accept the deserialized objects without having to manage the AMQ
connection.

Change-Id: I5604664e1ecadc41b25ef201e0a5384c6a308f5a