DAQ-3329 GDA client throws NPE on exit.
Some clients may not delegate Spring to start @Component objects.
In this case SpringApplicationContextFacade does not contains a valid
applicationContext which, on cascade may generate the NPE as in this
issue. That problem is fixed (see
SpringApplicationContextFacade.removeApplicationListener).

The analysis of the code revealed also an error in a related action
implementation
(SpringApplicationcontextProxy.addDisposableApplicationListener)

Change-Id: I1b1315493c59669ebd0d5dd66f5124bed461b660