DAQ-3290 Add Tycho config to build Javadocs for GDA Server
This uses the tycho-document-bundle plugin. Regular maven dependencies
of the documentation module are recursively resolved and Javadoc is
generated for all exported packages in the dependency tree.

The execution is enabled via the 'javadoc' profile. A profile was used
as the plugin doesn't seem to have a skip property. Therefore to
invoke a build including Javadocs the command would be:

 mvn -DskipTests -Pgda,server,javadoc clean verify

At the moment this build takes over two hours (on a fast SSD) due to
a inefficiency in Tycho's JavadocMojo. This has been fixed for the
next release (see https://bugs.eclipse.org/bugs/show_bug.cgi?id=564168).

Change-Id: Ifd1015b4b08df28cc7d12c6689c3aa6507881638