throw FactoryException if timerName not set
instead of calling Finder.find(null). This causes a ClassCastException
in unit tests if any findables don't have a name set (because they are
found by type).

Do not call super.configure from DummyCounterTimer as this will throw
an exception if the timer is not set, but DummyCounterTimer sets its
own timer.

Change-Id: I8fd8c99dfc9b02d44636191d0305fa821ae295db