K11-1130 After a diffraction the stream disappears
The bug was caused by incompatibility between the double axes in the
streaming and the MapPlotManager which is central for the original
mapping perspective.
The solution has been to avoid the double axes and display only the
camera pixels on the axes. At save or run the point/line/grid
coordinates are transformed according to the pco cameraToBeamMap
(ScanningAcquisitionController.transformPixelsToBeamDrivers). At load
coordinates are transformed according to the pco cameraToBeamMap
(ScanningAcquisitionController.transformBeamDriversToPixels)

To implement what described above, the CameraToBeamMap class has been
changed so to define a vector instead of set of "named" properties.
Most of the changes out of ScanningAcquisitionController are related to
this.


Change-Id: Ic1781cfb0a82e524d291574b18c679fcf29730c2