Use DS to obtain SessionService instead of other plugin's ServiceHolder
This prevents NPE when the GDA server is started:

 !MESSAGE bundle uk.ac.diamond.daq.messaging:1.0.0.qualifier
 (384)[JsonMessagingServiceImpl(93)] : The activate method has thrown an
 exception

Using the ServiceHolder from the core plugin is not ideal as this
service component can be activated before ServiceHolder itself is.

We then get an NPE when JsonMessagingService is activated as
ServiceHolder.getSessionService() returns null.

This didn't actually break anything as it seems that DS/SCR retries
activating the component later on after ServiceHolder has been injected
with its referenced services.

Change-Id: Ice5cc2073d700232ea2325be460383cb216f4a9e