DAQ-3272 Delete old versions of GDA_{Developer,User}_Guide
These manuals were extracted from this repo, converted to markdown,
and added to the gda-documentation repository.

These old copies are redundant and potentially misleading at the
moment. It has yet to be decided if eventually the content from
gda-documentation will be moved back into gda-core (enabling docs to
be close to source).

Change-Id: I97e0a140c1a6deffee36b50c8249a4c98dafdfb7