DAQ-3340 Ensure XANES scan model is NOT continuous
The constructor for AxialMultiStepModel used by the XANES GUI is the
only one that doesn't set continuous = false, so we need to do it
explicitly.

If continuous is set true, the model validation will fail (because the
gaps between the ranges are too large) and the status panel will not be
updated correctly.

Change-Id: I00b122a2b8d3c5ed6d2510f9d039ddaa0a321259