DAQ-3314 Allow dark/flat field collection with any detector
Replace the single "snapshot detector" with a map that allows us to
specify a detector to use for the snapshot corresponding to each "main
scan" detector.

This can for example be used to specify, for each Malcolm scan, a
Malcolm device that uses the same detector as the main scan but does not
open the shutter.

Change-Id: I522c4d0bd2d3ef24cccb6702934cd228d1ca185e