Add SpringProxyParser to GDA Spring handlers
This allows beans to be imported (and optionally registered as services)
with slightly less boiler plate. It still uses the GdaRmiProxy that was
as before but wraps it in a custom element to make it explicit that
it was creating a proxy rather than defining a new bean.

Where before a bean could be imported and registered as

    <bean id="bean_name" class="uk.ac.gda.remoting.client.GdaRmiProxy" />
    <bean class="gda.util.osgi.OSGiServiceRegister">
        <property name="class" value="gda.commandqueue.Processor" />
        <property name="service" ref="bean_name" />
    </bean>

This can now be reduced to

    <gda:proxy id="bean_name">
        <gda:service interface="gda.commandqueue.Processor" />
    </gda:proxy>

If no services need to be registered this can be reduced to

    <gda:proxy id="bean_name"/>

Change-Id: I69186bc3e027398bdefd573ce1e3843c0fb1d942