B18-477 : Spreadsheet view : duplicating values across rows picks up incorrect column
This fixes a bug where attempting to duplicate a value in the spreadsheet view
across multiple rows leads to values being duplicated in the incorrect column
(i.e. not the one right-clicked on).
This was caused by a NPE when selecting a filename cell value, leading
to the 'currently selected column' not getting updated correctly and being
out of date. Values are then duplicated for the last 'good' column selected
(e.g. motor position), rather than the selected file name.

Change-Id: I4424313128eb4190ed6f159b4252f3f67b18d3a4