I07-344 Add hkl writing to ProcessingMalcolmDevice
The hkl position datasets are required per point. For Malcolm this is
tricky as everything per point has to be written by Malcolm (i.e. by
the PandA or detector).

The hkl positions are provided by diffcalc so HklAdapter provides
abstract methods to be implemented in a Jython object which can involve
the diffcalc calls.

This is written once (except nested scans) per scan using the set
positions from the point generator at the start of the scan. Ideally
the PandA readbacks would be used but this requires additional
complexity.

Change-Id: I8dd3b433f1f2e97b9441452836cebed48f3f85d3