DAQ-3373 Configure CalibratedAxesProvider consistently
Firstly we restore the call to LiveStreamConnection::connect.

Secondly we ensure that the CalibratedAxesProvider in the live stream
connection gets the stream shape injected early on so that it does not
return null datasets. This fixes a race condition where plotting objects
call getAxes() on the stream before the first frame arrives and sets up
data shape.

Change-Id: Ie89d9ad9947571da4b62c7cad08dfff84baf4266