DAQ-3385 Create GDAClassLoaderService
The idea is to have a central class loading service which can be
provided to libraries (Jython, Spring, etc.) instead of the current
mixture of solutions (Eclipse buddying, OSGi fragments,
DynamicImport-Package).

The approach here uses DynamicImport-Package to allow all classes to
be loaded provided they are in exported packages from their containing
plugin.

Change-Id: I6364dec919b1a08359405939e235249e3fccd62d