K11-1147 Apply NexusTemplate to GDA Nexus scan files
At the core of the change is to add the ability to apply a nexus
template to any GDA nexus file.
The  client properties defines a set of
AcquisitionConfigurationProperties which defines, at client level, a
specific acquisition type, consequently, is natural to add here a list
of Nexus Templates URL.
At the ScanningAcquisition model level, to keep tis model consistent,
the request to apply the Nexus Templates is handled as a
ProcessingRequest (Key + Object) and only later, in the server, in the
ScanRequestFactory, converted to a standard ScanRequest.

- Add nexusTemplates property to AcquisitionConfigurationProperties
- Updates unit test in AcquistionConfigurationTest and
common_instance_java.properties
- Defines ApplyNexusTemplatesRequest (client), ApplyNexusTemplateHandler
(server) and update DocumentMapper
- Implements tests ApplyNexusTemplateTest
- Updates the ScanningAcquisitionController (client) and the
ScanRequestFactory (server)
- Defines unit tests for ApplyNexusTemplate in ScanRequestFactory
ScanRequestFactoryTest, ScanRequestFactoryTestConfiguration

Change-Id: I9742e827d776be67d0d567e229217a8b238882f1