DAQ-3279 Decouple ConcurrentScan's Json messages from old style events
This is required to ensure that the Json messages sent during a scan
are compatible with expectations of live processing and visualisation
consumers.

* The scan STARTED message must contain the filepath
* The UPDATED messages should only be sent after data is written to file
  otherwise consumers might not be able to read it

Due to the design of these classes (e.g. nested scans and async file
writing) using the existing events do not align with the accepted
Json messages. This could potentially be changed but would be non
trivial.

Change-Id: I844e7be99338ed6b7d46a339229a719032163076