I11-467 Ensure raw cvscan data is written
Raw data should be written when the cvscan enters REDUCTION state.
In some circumstances, this event is not received and raw data would
not be written.

This adds a second chance to write the raw data but includes a flag
to prevent it being written twice. This is not an ideal workaround but
should reduce the chance of of data being lost.

Change-Id: I2480f305a77377011b8721b39ac13f3600c93f80