[DAQ-3347] use a GroupMetadataNode in AbstractNexusMetadataDevice
This removes the duplication by treating the root node the same as
other nodes.
Note that this change removes the distinction between
predetermined fields (those with an explicit setter) and custom fields
(those without) as GroupMetadataNode does not know about the setters
and so it felt wrong that it should separate the two.

Change-Id: I6cde273b896d193bc11be7c401c3723cb9cbe8a0