[DAQ-3347] add NexusMetadataDevice as a generic nexus metadata device
AbstractMetadataNexusDevice previously existed as a place to put the
common behaviour of the specific metadata devices such as
SourceNexusDevice, MonochromatorNexusDevice, etc. However, with the
introduction of a MetadataFields and the ability to have custom fields,
it may be useful to have a completely generic subclass that makes this
behaviour available generically.

Change-Id: Iacd5eb15e84470016e84f460211c759f38a73149