K11-990 Remove deprecated methods from SlimPositionerComposite
It was not possible to replace ClientSWTElement.createButton with a more
recent ClientSWTElement.createClientButton because of incompatibility of
the actual layout with the GridLayout used in createClientButton.

As ClientSWTElement.createButton is deprecated and removing it from
SlimPositionerComposite would allow to remove it from ClientSWTElement
too.

Consequently I
- copied ClientSWTElement.createButton in the SlimPositionerComposite
with minor changes
- deleted ClientSWTElement.createButton

At the same time I partially cleaned ClientSWTElement from deprecated
and not used methods.


Change-Id: I7526cb3547a66ac70265eca5f758e029c1f09978