DAQ-3400 Improve LDAP when servers are unavailable and general cleanup of LdapMixin
Previously a list of servers was constructed (from property or DNS
resolution) which was tried in turn to create an LDAP context however
if a server was unavailable it would potentially block for a long period
of time.

Instead pass the list to the context environment as a string separated
list as documented at
https://docs.oracle.com/javase/jndi/tutorial/ldap/misc/url.html
Additionally add connection/read timeouts to the LDAP context
environment so they do not block forever. Make these timeouts settable
via a GDA property in case of network issues etc.
Reorder methods and tidy code in LdapMixin for improved readability.

Change-Id: I20a71829bb794eaf1c23243b9428ec61ae83ed4d