[DAQ-3368] rewrite NexusDetectorNexusWriter from scratch
The previous implementation is a copy/paste from the
NexusDataWriter.writeHere method, changed as little as possible in
minimize the likelihood of breaking anything, as this is quite
complex code. However, this was overly complex code badly in need
of refactoring or rewriting. Having understood what the code was doing
I felt confident to completely rewrite it.

Change-Id: I4ce725b6cff3e4400fc5c759317dde36731e6954