DAQ-3390 Extract logic for mapping scan region shape onto scan path model in mapping perspective from PathInfoCalculatorJob
Extract logic for mapping scan region shape onto scan path model in
mapping perspective from PathInfoCalculatorJob because the logic in
question is independent of calculating scan points.

The broader aim of this change is to facilitate the drop-in replacement
of ScanPathInfoCalculatorJob with another equivalent class in order to
facilitate plotting via ScanSpec (see DAQ-3390).

Change-Id: I4b9eea6d6c4288608a48fac573b4a0a7025e3108