DAQ-3423 Disable Jython bytecode generation
As of Jython 2.7.2 the bytecode $py.class files are created with
restricted permissions (u=rw only) and these are written alongside the
source (exported exploded plugin directory in the client case). If a
client is subsequently opened by another user, Jython fails as the
compiled code can't be read.

This interpreter manager is called from both the server and the client
for scan point generation. For the server this will share/augment the
existing PySystemState which should be configured appropriately
(GDAJythonInterpreter).

Change-Id: I6ece291554370a61a1f440df18147e229608e4a5