K11-1175 Processing Request GUI component for DIAD
- ProcessingRequestBuilder allows ProcessingRequestKey(see below) to
create a collection of builders for classes extending
ProcessingRequestPair
- Creates ProcessingRequestContext and ProcessingRequestKey to support
ProcessingRequestComposite(see below)
- Refactors SelectFileComposite so may returns not only a set buttons
but also a ToolItem (used in the ProcessingRequestComposite file
selection)
- Creates ProcessingRequestComposite to add and configure processing
requests. Each row is implemented as ProcessingRequestRow.
- Refactor SelectableContainedCompositeFactory so that each time the
innerContainer is repainted, the scrolledInnerContainer is resized
correctly. this is relevant in this change so adding or removing rows
does not push out other composite elements. This update is a major
advantage for any future innerContainer.



Change-Id: I127d8593c8df3645638c5bbd2aad5dcaa792bb52