K11-1212 Add system to prepare scan requests according to scan type
The rationale behind this mechanism is that while ScanRequests contain
parameters which will vary from scan to scan, it also contains data
which depends on the scan type e.g. for a diffraction scan the beam
selector should be in the diffraction beam position (though currently
this is handled by Malcolm).

An initial DiffractionPreparer is implemented which simply adds a per-
scan monitor if it is given via a local property. A corresponding config
change defines this monitor as a radiograph frame collector.

This change is given as is because it is needed fairly soon. However
while the overall mechanism may have merit, it is clear that the imple-
mentation of DiffractionPreparer is not related to generic diffraction
scans but to a specific K11 problem and solution via a per-scan monitor.

Change-Id: I7b613797b028bc08764f61f8aa63f323d6779b85