DAQ-3386 Use GDAClassLoaderService for Jython class loading
Most of the work in ApplicationEnvironment is no longer required (or was
unused beforehand) the extension point lookup has been moved to
GDAJythonScriptApi to be close to GDAJythonInterpreter.

The new class loading service is added to the common feature so that it
is included in the client and server products.

Change-Id: Iccb19ecf5574b7b9fbc9f05077c8548348ba6a57