[DAQ-3463] refactor NXdetector method to add external links
add methods to take the rank of the external dataset as an additional
parameter, with existing methods delegating to the new methods so as
not to break existing code.

Deprecate method that take isPointDependent or isDetectorEntryData as
these boolean flags are ignored by NexusDataWriter and isPointDependent
is ignored by NexusScanDataWriter also.

Refactor addScanFileLink to call addExternalFileLink to reduce
duplication

Improve Javadoc.

Change-Id: I2185618c449e7240613f2a4548ea03f1e3676db9