I14-857 Allow lines to track to be unset in XANES submit section
It is a requirement to run XANES scan without lines to track set,
so allow it but check to avoid NPEs.

Note that selecting a line, then "unselecting" it (i.e. selecting the
blank line at the start of the list) causes a LineToTrack entry to be
created containing an unmodifiable Collection that causes problems later
on in deserialisation. This fix identifies this case and sets the entry
to null, which is the value you get if you never select a line.

Change-Id: Ib53f509698cc7a238c191fc842c761f0c1aa2157