DAQ-3478 Add all bundles providing split packages to sp resolver
As the split package resolver is added to the client, the bundles have
been marked as optional.

This creates a potential risk that a package may be exported from this
resolver but not actually contain any classes if the bundles were all
not present.

To ensure this will not be a problem, make sure that all possible split
package providers are required by the resolver.

This doesn't actually "resolve" any more split packages yet but if
issues are encountered the corresponding packages can be added to the
exports of the resolver.

Change-Id: Iae839cf461acccc6a2705d22f22faf69111ad3a9