I07-354 Improvements and bug fixes for Odin Excalibur
Mostly renaming variables and adapting class after the addition of a
second Odin data writer to I07.

There were also channel access issues reported (I07-354) which this
hopes to address.

The external file link is also switched to use addExternalFileLink
so that an NXdata is created when using the NexusScanDataWriter.

Change-Id: I9f1af8203fd8a54fa776ae28f3f0593780559c19