Add PyTestRunner to run tests in the GDA environment
Using Pytest allows beamlines to write automated acceptance tests
that can use the test discovery and reporting features of
a well known framework as well as allow the use of fixtures
for common tasks that require setup/teardown.

Markers and decorators allow tests to be parameterized and
skipped according to GDA configuration states, eg "only run
this test if the server was started with this profile".

This class acts as a Python environment context wrapping, the standard
stdout so that pytest can use it as it would for a tty.

Change-Id: Icac135661702389a7d52c7d6f94c32b1ee82ffa3