DAQ-3509 Fix Solstice scanning class loading for Jython
After switching Jython in the GDA server to use the class loading
service the initialisation for Jython in Solstice scanning was causing
Jython to break as the class loader in the existing PySystemState was
replaced by a CompositeClassLoader which did not contain the
GDAClassLoader.

Now ensure that the class loader is only replaced if Jython has not
previously been initialised (by GDAJythonInterpreter).

Change-Id: I0764f10d8ce20a705495e808837b0c6e34c328de