I22-1078 Throw expception if no Tetramm file is written
If the Tetramms are configured for Panda/mapping acquisitions,
data collections using the TFG require them to be reconfigured. If
they have not been reconfigured, no frames are received by the HDF
plugin and the (lazy open) file is not created.

This is almost certainly an error and aborting the scan makes it
clear to users that something is wrong.

Change-Id: Ibf5c384bad516046d1fdfd16cb1c06f797c5ed77