K11-1245 Extract base class from CalibrationFrameCollector
Base class FrameCollectingScannable performs an AcquireRequest before
our main scan and copies its dataset into our main scan file.
A DetectorDocument must be set to specify the name of the detector for
the AcquireRequest and the exposure, as well as the name of the inner
detector if using Malcolm.
This will enable frame collection with detector parameters different
to the main scan, as required by the linked ticket.

Change-Id: Ied7a0ccbd6f236e521e98cc44879c78362d74622