DAQ-3597 Move Jython package manager initialisation to static block
Initialisation of the PackageManager with Java classes/jars only
needs to happen once in the lifetime of the GDA server.

In the previous state it was called for every reset_namespace() which
causes the static lists inside the manager to extend in length with
duplicated entries.

This eventually causes a complete failure of Jython/GDA as the server
process becomes stuck with "Too many open files".

Change-Id: Icb05e61f38df80f28c976a91caa31258608a359d