Prevent NPE when retrieving info for unknown clients
If a client is left open after the server is restarted, the client's
UID is not recognised by the server. If it then tries to query its
details, a NPE is thrown on the new server.

With clients now displaying the source of input for terminal commands,
details were queried much more often.

Change-Id: I6c32a640bf4687b3031ab5340538864b1e8257f7