I18-479 Fix and complete implementation of ExperimentPerspectiveListener
This listener is a hack to treat the editors required for the Experiment
Perspective as views i.e. bound to that perspective only. This stopped
working correctly some time ago (IPerspectiveListener3 and 4 methods are
not called: https://bugs.eclipse.org/bugs/show_bug.cgi?id=408309) so a
second listener was added in ExperimentExperimentView [sic]. This last
listener is removed in this change.

The listener has been reimplemented using OSGi events. This was the only
way I could reliably get a reference to the perspective we are switching
*from*. Additionally, the listener now restores editors that were open
before the ExperimentPerspective was switched to.

Change-Id: I0eb8d857ff857b87ec2716328c483f88c5c508a4