[DAQ-3567] ensure correct tagging of NXdata group for hkl scans
Each of the hkl fields should be added to the NXdata. The default
axis field (in the @axes attribute) should be the one that has the
larger range. To get this information we need to add the ScanObjects
to the IScanDataPoint, and then set the appropriate ScanObject in the
ScannableNexusDevice.

Change-Id: Ie2449c93564a3795a4e02cd6dc59a4ca5ed0ab4e