[DAQ-3635] creating nexus file earlier to add all per-scan monitors to annotation manager
Previously we weren't adding all per-scan monitors to the annotation
manager, as those that are configured by the NexusScanDataWriterConfiguration
and the dependencies between scannables are not added until we create
the nexus file. This mean that they weren't added to the
AnnotationManager, resulting in their @ScanFinally methods not being
called, and therefore the nexus datasets hanging around between scans.

Change-Id: If3aa2e78b5d289f62a98df7acb5891e8afd9e67d