Enable spring completion and validation for gda namespace
Eclipse IDE now uses Wild Web Developer for XML validation. This catalog
can be added to Lemminx to allow the resources to be found.

While it could be automatically set as a preference by Oomph, for now it
can be enabled by adding gda-spring-catalog.xml from
  gda-core.git/uk.ac.gda.core/src/gda/spring/namespaces/gda
to Windows > Preferences > XML (Wild Web Developer) > Catalogs
and also adding it to Windows > Preferences > XML > XML Catalog using a
"Delegate catalogue" with a "System ID" matching
  http://www.diamond.ac.uk/schema/gda/gda

Change-Id: Ifad62dc1d28d6d4c446410c414ecbf03e1df510d