Add allOfType spring parser
This allows properties to be autowired without using component scans.
It is intended to replace SPEL expressions that query the context for
beans of a given type.  This is particularly useful if the collection
being set as a property depends on which profiles are active at runtime.

eg a SPEL expression such as

    <property
        name="foo"
	value="#{getBeanFactory().getBeansOfType(Class.forName('com.example.Foo').values())}"
    />

can be replaced with

    <property name="foo">
        <gda:allOfType type="com.example.Foo"/>
    </property>

See DAQ-3650

Change-Id: I32a516d79583bc0cada9b7bc7c54bcefa4c9d996