DAQ-3651 Fix DeviceInterceptor class loading
Since the class loader on the Enhancer was changed to be the dynamic
loader (which can only see exported classes) the proxy creation could
fail if reflection methods (e.g. getDeclaredMethods()) resulted in
the need to load types which are not exported by the bundle that the
wrapped type lives in.

This is fixed by obtaining a class loader with its parent set to the
loader of the class being wrapped. It still requires the dynamic loader
as a way to see the Spring classes required for the proxy.

Change-Id: Iafd9267ee134e8da4155b9e3b5e7926378f8fe0d