K11-1347 Frame Capture sets PCO camera exposure to 0.0
In DIAD the client submits a SCanningACquisitin to the server.
Then, on the server, the ScanningAcquisitionFactory, creates and
populate a ScanRequest. In this process it creates also the
ProcessingRequests using a collection of ProcessingRequestHandler.
In particular the FrameCaptureRequestHandler fills the
FrameCollectingScannable with the right DetectorDocument. This change
avoids the CalibrationFrameCollector.configureCollection to overwrite an
already set (by the FrameCaptureRequestHandler) DetectorDocument.

- implements FrameCollectingScannable.getDetectorDocument()
- updates CalibrationFrameCollector.configureCollection

Change-Id: I0567f2d61e96baa7c386d778962dd3e9ae99973b