DAQ-3391 Make ScanBuilder aware of non default NXEntry name
Take account of mapping NXEntry name property (since load function is
only available for mapping generated files), defaulting to 'entry' if
property is not defined or cannot be matched in specified file, since it
may have been written when the property was not set.

If even the default is not successful, an empty Optional is returned (as
before) and a warning is added to the log.

Change-Id: I987c8ebfaa115ca4cf65a2498e4401785bfd10e5