i14, i18: Disable further mouse wheel events on Mapping
Once again, this is because the controls in question sit on a vertically
scrollable view and users often try scrolling with mouse wheel but the
focus goes on these controls (here combo viewer, spinners) and wheel
changes the selection on them instead.

Some minor tidying: replaced deprecated classes and defined local simple
variables with var.

Change-Id: I862df80076a252c11afccdeab30e66830c7c9ca5