K11-1368 Refactor create new AcquisitionDocument
At the moment both the Diffraction and the Tomography use their own
instance of a ScanningController. However this is redundant as the only
difference is the controller internal ScanningAcquisition. Consequently
is possible to delegate Spring to instantiate the AcquisitionController
while leaving to the underlying client composite the duty to set its
acquisition configuration document.

To achieve that:

- Refactor the ScanningAcquisitionController from a Spring prototype to
a simple bean, consequently has been possible to refactor the
ClientSpringContext, which is a bean on its own, to instantiate an
AcquisitionController.

- Implements AcquisitionKeys to stress the unique relation in the GUI
between the AcquisitionPropertyType and AcquisitionTemplateType

- Refactored the DocumentFactory and ButtonControlledCompositeTemplate
to use AcquisitionKeys

- Moves the AcquisitionController interface in gda.client being a purely
client object

- Because the AcquisitionController delegate the DocumentFactory to
create a new document, and because the Controller is client wide
available through the ClientSpringContext, is possible to introduce a
new method newScanningAcquisition(AcquisitionKeys acquisitionKey) and
remove both setDefaultNewAcquisitionSupplier(Supplier<T>
newAcquisitionSupplier) and createNewAcquisition().
AcquisitionConfigurationView has been refactored accordingly.

- Updates SelectableContainedCompositeFactory so that the active
NamedCompositeFactory is properly disposed when the perspective change

- Updates ButtonControlledCompositeTemplate so that an acquisition with
no defined properties returns an empty configuration view (can be
improved later) with a warning message (see BeamSelector scan in
Diffraction)

- Moves in gda.client AcquisitionConfigurationException and
AcquisitionControllerException.



Change-Id: I706599719e9c999962fe0755566ff291502e98dc