[DAQ-3619] fix malcolm detector exposure time written as 0.0
When configuring a malcolm device, an exposure time of 0.0 for a
detector is the default value. The value is interpreted as meaning
'use the maximum value', i.e. if framesPerStep is 1, then this will be
the same as the exposure time of the malcolm device.

When calling validate or configure on the actual malcolm device with an
EpicsMalcolmModel, the reply from the malcolm device includes the
actual exposure time that will be used for the malcolm device.

Prior to this change, when 0.0 is the configured exposure time for a
malcolm controlled detector, this is the value written into the nexus
file, as the value of the field NXdetector.count_time instead of the
actual exposure time.

Change-Id: I4bb1681a09f7ce1f4c1daf8f825e950a35771cf3