Revert change to use alias instead of vararg_alias for some builtins
While vararg_alias should not be needed, there is a bug in how Jython
handles vararg methods that requires the boxing to be done in GDA (via
the translator). See https://github.com/jython/jython/issues/100

This will reintroduce some of the parsing faults that the previous
change intended to fix but they are less prevalent than the ones the fix
caused.

Main example that prompted this is trying to set the position of a
scannable that takes multiple values.

    >>> pos base (1, 2)

gets translated to `pos(base, (1, 2))` correctly but Jython fails to
convert the arguments to match the vararg signature and a TypeError is
raised. This leads to the help for `pos` being shown to the user for
what would otherwise be valid commands.

Change-Id: Ib06ac315cb05ba2bc527cba7f03e01650b3cd84a