K11-1382 Ensure exposure changes are picked up in frame capture requests
This change fleshes out the implementation of a frame capture requests
and fixes a couple of bugs along the way.

ProcessingRequestProperties now has a new field FrameCaptureProperties
to configure via properties the malcolm & detector names used for
capture requests associated to a particular acquisition type.

The request now operates with its own FrameRequestDocument document type
instead of the more generic DetectorDocument.

ScanningAcquisitionDetectorUpdater is simplified and fixed to ensure
changes in exposure are injected into the relevant detector and frame
request documents.

Change-Id: I24801d8214655f9a2ff323fc8d1879f3d4ad9003