Replace draw2d ScaledGraphics with internal version
To prevent scaling bugs around drawing 0 width graphics. Original
draw2d implementation included checks that the zoomed size of some
graphics was non-zero but not others. In order to add the checks, the
whole class needs to be duplicated as required methods are private.

This change is not ideal and it would be better to migrate to the
better supported gef framework but for now this solves the UI problems
seen on MX beamlines.

See I19-538

Change-Id: I24db0a568eb48e185b858c234ff3f70d96e17e8e