DAQ-947 Add multiline commands to JythonShell
If a command includes linebreaks, treat it as a series of commands and
run in exec mode instead of single mode.

The main change to existing behaviour is related to the displayhook
mechanism. In `single` mode, any unbound object is printed and the last
of these is assigned to the `_` variable, eg

    >>> if True:
    ...     "helloWorld"
    ...
    'helloWorld'
    >>> print _
    helloWorld

With the new approach, single line commands will maintain this behaviour
but multiline commands will be treated as if they were scripts where the
displayhook is not used.

    >>> 2 + 3
    5
    >>> print _
    5
    >>> if True:
    ...     "helloWorld"
    ...
    >>> print _
    5

The additional functionality gained from this change is in allowing
multiple statements to be edited (and run) together. This can be
particularly useful when copying/pasting series of commands from
elsewhere or repeatedly running a series of commands required to
configure devices during development.

    >>> def demo(a, b):
    ...     return a + b
    ... print demo(2, 3)
    ... print demo("hello", "World")
    5
    helloWorld

Once a command has multiple lines, it will be treated as incomplete
until it ends with two consecutive new lines. Ctrl-n is now available in
the shell to add a new line to an existing command if required to add to
an already complete command.

This does not affect the JythonTerminalView in the RCP client.

Change-Id: I194c84c56517d9ffb8ce1defd9a68788aff89939